requirements.txt
drawsvg
imageio
kornia
numpy
opencv-python
Pillow
pyopencl
shapely
matplotlib
scikit-image
tqdm
